+++ src/weather/weather.ts
@@ -1,5 +1,6 @@
 import {getBooleanSetting, getStringSetting} from '../settings';
 import {allWeatherNames, getWeatherSettings, WeatherEffectName} from './data';
+import {isFirstGm} from '../utils';
 
 export function dayHasChanged(game: Game, deltaInSeconds: number): boolean {
     const now = game.pf2e.worldClock.worldTime;
@@ -33,44 +34,59 @@ async function applyWeatherSound(game: Game, effectName: WeatherEffectName): Pro
     }
 }
 
-async function updateSceneWeather(scene: Scene | null | undefined, eff: WeatherEffectName): Promise<void> {
+function getSceneWeather(game: Game, scene: Scene | null | undefined): WeatherEffectName | null {
     if (scene && getWeatherSettings(scene).syncWeather) {
-        console.info(`Setting scene ${scene.name} weather to ${eff}`);
-        if (eff !== 'sunny' && eff !== '' && allWeatherNames.includes(eff)) {
-            await scene.update({'weather': eff});
-        } else {
-            await scene?.update({'weather': ''});
+        const effectName = getBooleanSetting(game, 'enableSheltered')
+            ? ''
+            : getStringSetting(game, 'currentWeatherFx') as WeatherEffectName;
+        if (effectName === 'sunny') {
+            return '';
+        } else if (allWeatherNames.includes(effectName)) {
+            return effectName;
         }
+    }
+    return null;
+}
+
+function getWeatherPlaylist(game: Game, activeScene: Scene | undefined | null): WeatherEffectName {
+    if (activeScene && getWeatherSettings(activeScene).syncWeatherPlaylist) {
+        return activeScene.weather || (getBooleanSetting(game, 'enableSheltered') ? '' : 'sunny');
+    } else {
+        return '';
+    }
+}
+
+async function syncSceneWeather(game: Game, scene: Scene | null | undefined): Promise<void> {
+    const effectName = await getSceneWeather(game, scene);
+    if (scene && effectName !== null) {
+        console.info(`Setting scene ${scene.name} weather to '${effectName}'`);
+        await scene?.update({'weather': effectName});
+    }
+}
+
+async function syncScenePlaylist(game: Game, scene: Scene | null | undefined): Promise<void> {
+    const weatherEffect = getWeatherPlaylist(game, scene);
+    console.info(`Setting weather playlist to to ${weatherEffect}`);
+    await applyWeatherSound(game, weatherEffect);
+}
+
+async function syncWeather(game: Game): Promise<void> {
+    if (isFirstGm(game) && getBooleanSetting(game, 'enableWeather')) {
+        const currentScene = game.scenes?.current;
+        const activeScene = game.scenes?.active;
+        if (activeScene && currentScene && activeScene.id === currentScene.id) {
+            await syncSceneWeather(game, activeScene);
+        } else {
+            await syncSceneWeather(game, currentScene);
+            await syncSceneWeather(game, activeScene);
+        }
+        await syncScenePlaylist(game, activeScene);
     }
 }
 
 export async function setWeather(game: Game, effectName: WeatherEffectName): Promise<void> {
-    if (getBooleanSetting(game, 'enableWeather')) {
-        // always persist the current one without checking for turned off weather
-        await game.settings.set('pf2e-kingmaker-tools', 'currentWeatherFx', effectName);
-        // fall back to sunny if weather is disabled
-        const eff = getBooleanSetting(game, 'enableSheltered') ? '' : effectName;
-        const currentScene = game.scenes?.current;
-        const activeScene = game.scenes?.active;
-        await updateSceneWeather(currentScene, eff);
-        await updateSceneWeather(activeScene, eff);
-        if (activeScene) {
-            if (getWeatherSettings(activeScene).syncWeatherPlaylist) {
-                const weatherEffect = activeScene.weather
-                    || (getBooleanSetting(game, 'enableSheltered') ? '' : 'sunny');
-                console.info(`Setting weather playlist to to ${weatherEffect}`);
-                await applyWeatherSound(game, weatherEffect);
-            } else {
-                console.info('Disabling weather playlist since syncing was turned off');
-                await applyWeatherSound(game, '');
-            }
-        }
-    }
-}
-
-export async function syncWeather(game: Game): Promise<void> {
-    const weather = getStringSetting(game, 'currentWeatherFx') as WeatherEffectName;
-    await setWeather(game, weather);
+    await game.settings.set('pf2e-kingmaker-tools', 'currentWeatherFx', effectName);
+    await syncWeather(game);
 }
 
 export async function toggleSheltered(game: Game): Promise<void> {
@@ -83,5 +99,34 @@ export async function toggleWeather(game: Game): Promise<void> {
     const isEnabled = !getBooleanSetting(game, 'enableWeather');
     await game.settings.set('pf2e-kingmaker-tools', 'enableWeather', isEnabled);
     await syncWeather(game);
+}
+
+export async function onPreUpdateScene(game: Game, scene: Scene, update: Partial<Scene>): Promise<void> {
+    if (isFirstGm(game) && getBooleanSetting(game, 'enableWeather') && update?.active) {
+        // make sure to set the weather effect when viewing a scene and then activating it
+        const weather = getSceneWeather(game, scene);
+        if (weather !== null) {
+            update.weather = weather;
+        }
+    }
+}
+
+export async function onUpdateScene(game: Game, scene: Scene, update: Partial<Scene>): Promise<void> {
+    if (isFirstGm(game) && getBooleanSetting(game, 'enableWeather') && update?.active) {
+        // activating a scene plays its weather playlist
+        // setting the correct scene weather is done pre update or on render
+        await syncScenePlaylist(game, scene);
+    }
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export async function onRenderScene(game: Game): Promise<void> {
+    // disabled because this produces foundry race conditions
+    // const current = game.scenes?.current;
+    // if (isFirstGm(game) && getBooleanSetting(game, 'enableWeather') && current) {
+    //     // when changing scenes, also ensure that the weather syncs without activating the scene
+    //     const current = game.scenes?.current;
+    //     await syncSceneWeather(game, current);
+    // }
 }